Quadrants
I finished the basics of the Quadrant class as well as a basic skeleton of the Body class. The Quadrant class is used to represent quadrants in the 2D space. Each quadrant also has the ability to create sub-quadrants of that quadrant. The Body class is going the be used to represent each individual bodys in the simulation. I also created a class called Point which I might want to use in order to pass (x, y) points.