Finished Bodies and QuadTrees
I finished writing both the Body and the QuadTree classes. The Body class now contains all the information to represent bodies in 2D space. It also has methods to update its position and add the force enacted upon it by another body. The QuadTree class is a data structure that holds each quadrant in a node, where each node in the QuadTree is a new instance of the QuadTree class. The QuadTree class has methods to add new bodies to the tree as well as approximate the forces the bodies in the QuadTree enacts upon another body.