Basic Functionality
I have completed the basic NBody simulation. The algorithm used to calculate forces is in place as well as graphical component. Currently the simulation only creates stationary bodies uniformly in a circle and runs it from there. I also found some problems with the QuadTree and Body classes and fixed them.