Circular Orbits
I added some basic physics to give each body the correct x and y velocities to achieve a circular orbit around the central mass. I also edited the body class so that I can make the central mass stationary. The reason for this is that if allowed to move, the central body will move off the screen.